refactor: restructure Panel components and move to PanelWrapper for better organization
