+++ src/components/PanelWrapper/TopbarTrigger.tsx
@@ -1,11 +1,11 @@
-import Panel from './Panel';
+import PanelWrapper from './PanelWrapper';
 
 const TopbarTrigger = () => {
   return (
     <div className="flex h-full gap-2">
       <span className="mx-8 w-px bg-neutral-800"></span>
       <div className="flex h-full justify-center">
-        <Panel />
+        <PanelWrapper />
       </div>
     </div>
   );
